+++ destructify/fields/base_field.py
@@ -20,8 +20,8 @@ class BaseFieldMixin(object):
         if self._take_attributes_from_base:
             if self.base_field.has_default and not self.has_default:
                 self.default = self.base_field.default
-            if self.base_field.has_convert and not self.has_convert:
-                self.convert = self.base_field.convert
+            if self.base_field.has_decoder and not self.has_decoder:
+                self.decoder = self.base_field.decoder
             if self.base_field.has_override and not self.has_override:
                 self.override = self.base_field.override